fix(heartbeat): cap configurable execution timeout
