+++ console/src/locales/ja.json
@@ -554,6 +554,7 @@
     "timeoutSeconds": "実行タイムアウト（秒）",
     "timeoutRequired": "必須",
     "timeoutMin": "1以上の値が必要です",
+    "timeoutMax": "3600以下の値が必要です",
     "unitMinutes": "分",
     "unitHours": "時間",
     "target": "返信先",
