+++ website/public/docs/heartbeat.en.md
@@ -86,7 +86,7 @@ source of truth for new changes.
 | **enabled**        | Heartbeat on/off. **Default false**; the schedule runs only when **true**.                                                                                                                                                                                                |
 | **every**          | How often: an interval string (`"30m"`, `"1h"`, `"2h30m"`, `"90s"`) **or** a space-separated **five-field cron** (minute hour day month weekday — same shape as cron jobs, e.g. daily 09:00: `"0 9 * * *"`). Cron is interpreted in the **process scheduler’s timezone**. |
 | **target**         | **main** — don’t send to a channel; **last** — send using **`last_dispatch`** for that agent; **inbox** — send to Inbox.                                                                                                                                                  |
-| **timeoutSeconds** | Maximum execution time for one heartbeat run, in seconds. **Default 300**.                                                                                                                                                                                                |
+| **timeoutSeconds** | Maximum execution time for one heartbeat run, in seconds. **Default 300**, valid range **1–3600**.                                                                                                                                                                        |
 | **activeHours**    | Optional daily window: `{ "start": "08:00", "end": "22:00" }`.                                                                                                                                                                                                            |
 
 If **every** is omitted, the built-in default applies (currently about **6
