+++ website/public/docs/heartbeat.zh.md
@@ -58,7 +58,7 @@
 | **enabled**        | 是否开启心跳。**默认 false**；只有为 **true** 时才会按计划执行。                                                                                                                                                      |
 | **every**          | 多久跑一次：间隔字符串（如 `"30m"`、`"1h"`、`"2h30m"`、`"90s"`），**或** 空格分隔的 **五段 Cron**（分 时 日 月 周几，与定时任务相同的五段写法，例如每天 9:00：`"0 9 * * *"`）。使用 Cron 时由进程内调度器的时区解析。 |
 | **target**         | **main** 不发送到频道；**last** 发到该智能体 `last_dispatch` 记录的频道；**inbox** 发送到收件箱中。                                                                                                                   |
-| **timeoutSeconds** | 单次心跳执行的最长时间，单位秒。**默认 300**。                                                                                                                                                                        |
+| **timeoutSeconds** | 单次心跳执行的最长时间，单位秒。**默认 300**，有效范围 **1–3600**。                                                                                                                                                    |
 | **activeHours**    | 可选，只在每天这段时间内跑：`{ "start": "08:00", "end": "22:00" }`。                                                                                                                                                  |
 
 `every` 未写时的默认间隔以程序内置为准（当前默认约为 **6 小时**，仍以你环境里的版本为准）。
